refactor: 유저 서비스 테스트 패키지명 변경
