+++ src/test/java/postman/bottler/user/application/BanServiceTest.java
@@ -1,4 +1,4 @@
-package postman.bottler.user.applications;
+package postman.bottler.user.application;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
