+++ src/test/java/postman/bottler/user/application/EmailServiceTest.java
@@ -1,0 +1,5 @@
+package postman.bottler.user.application;
+
+class EmailServiceTest {
+
+}
