+++ src/test/java/postman/bottler/user/application/KakaoServiceTest.java
@@ -1,0 +1,5 @@
+package postman.bottler.user.application;
+
+class KakaoServiceTest {
+
+}
