--- src/test/java/postman/bottler/user/applications/KakaoServiceTest.java
@@ -1,5 +1,0 @@
-package postman.bottler.user.applications;
-
-class KakaoServiceTest {
-
-}
